code refactor

+++ minsu3d/common_ops/src/bfs_cluster/bfs_cluster.cpp
@@ -14,7 +14,7 @@ All Rights Reserved 2020.
 // output start_len: (n, 2), int
 int ballquery_batch_p(at::Tensor xyz_tensor, at::Tensor batch_idxs_tensor, at::Tensor batch_offsets_tensor, at::Tensor idx_tensor, at::Tensor start_len_tensor, int n, int meanActive, float radius){
     const float *xyz = xyz_tensor.data_ptr<float>();
-    const int16_t *batch_idxs = batch_idxs_tensor.data_ptr<int16_t>();
+    const uint8_t *batch_idxs = batch_idxs_tensor.data_ptr<uint8_t>();
     const int *batch_offsets = batch_offsets_tensor.data_ptr<int>();
     int *idx = idx_tensor.data_ptr<int>();
     int *start_len = start_len_tensor.data_ptr<int>();
